move functions to reactjs module
