+++ packages/uxpin-merge-cli/src/framework/FrameworkModule.ts
@@ -1,5 +1,6 @@
 
 export interface FrameworkModule {
   getComponentMetadata:any;
-  PresetFileGenerator?:any;
+  PresetFileGenerator:any;
+  getLibraryBundleSource:any;
 }
