--- packages/uxpin-merge-cli/src/framework/common/index.ts
@@ -1,3 +1,0 @@
-import { FrameworkModule } from '../FrameworkModule';
-
-export const commonModule:FrameworkModule = {};
